+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Changes since v2.12
 - Automated browser testing has been improved in implementation and also in the coverage of the administration side
 - Form API create & edit requests are validated (#2098). This was meant to be added in 2.7 but the validation wasn't active by mistake.
 
+### Fixes
+- Various fixes in workflow editor UI
+
 ### Additions
 - The form administration pages now flag forms that have missing localizations. REMS also logs a warning on startup for these forms. (#2098)
 - There is now an API for querying and creating organizations #2039